Moved PointWeatherGateway. Removed dependency of InfoPanelPresenter on view.
