+++ src/main/java/dataaccessinterface/OkHttpsPointWeatherGatewayXml.java
@@ -1,4 +1,4 @@
-package dataaccessobjects;
+package dataaccessinterface;
 
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
